+++ sd-card/html/readconfigcommon.js
@@ -269,7 +269,7 @@ function MakeContrastImageZW(zw, _enhance, _domainname){
           return true;
      }
      else {
-          firework.launch("Device is busy, please try again when the round got completed!", 'warning', 10000);
+          firework.launch("Device is busy, please try again when the Digitalization Round got completed!", 'warning', 10000);
           return false;
      }
 }
@@ -297,7 +297,7 @@ function MakeRefZW(zw, _domainname){
           return true;
      }
      else {
-          firework.launch("Device is busy, please try again when the round got completed!", 'warning', 10000);
+          firework.launch("Device is busy, please try again when the Digitalization Round got completed!", 'warning', 10000);
           return false;
      }
 }